requirements.txt
flask
loguru
pillow
